Updating composer.json per email from Jordi Boggiano
---------- Forwarded message ----------
From: Jordi Boggiano <j.boggiano@seld.be>
Date: Thu, Jun 11, 2015 at 8:55 AM
Subject: Your Composer plugin requirements are potentially harmful
To:

Heya,

I am sending you this email because you have a Composer Plugin on
https://packagist.org that requires the `composer-plugin-api` in
version `*` or `1.0.0`.

Both of those are pretty dangerous:

- `*` would mean your plugin can get loaded in any future composer
including 2.0, and will probably break.
- `1.0.0` means that your plugin will not be loaded anymore as soon as
we add functionality and bump the plugin api to 1.1+

The solution is simple, open your composer.json and replace that require with:

	"composer-plugin-api": "^1.0"

Sorry for the trouble, the documentation had "1.0.0" in the plugin
example which mislead most plugin authors, but that is now fixed.

Cheers

--
Jordi Boggiano
@seldaek - http://nelm.io/jordi